# BUG: Index.take may add unnecessary freq attribute