``Index.freq`` should raise ``AttributeError`` in non-datetime-like index. But ``Index.take`` adds ``freq=None`` and broke the behavior. 

```
import pandas as pd
idx = pd.Index([1, 2, 3])

idx.freq
# AttributeError: 'Int64Index' object has no attribute 'freq'

idx.take([True, False, True]).freq
# None
```