@sinhrks 2015-08-12
Updated the release note. Pls review.